removed gpAssetService
